校验采样设置

+++ Torque/contracts/Sampling.cs
@@ -5,7 +5,7 @@ namespace Torque
     public class Sampling
     {
         public TimeSpan Time { get; set; }
-        public int Frequency { get; set; }
-        public TimeSpan Interval => TimeSpan.FromTicks(TimeSpan.TicksPerSecond / Frequency);
+        public double Frequency { get; set; }
+        public TimeSpan Interval => TimeSpan.FromTicks((long)(TimeSpan.TicksPerSecond / Frequency));
     }
 }
